requirements.txt
anyio==3.7.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.95.2
ffmpeg-python==0.2.0
future==0.18.3
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
psycopg2-binary==2.9.6
pydantic==1.10.8
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
